payment-gateway-service/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
protobuf
pydantic
python-dotenv
